+++ lib/models/potentialOperationsResult.ts
@@ -22,9 +22,12 @@ import { LiveValidationError } from "./liveValidationError"
  *
  */
 export class PotentialOperationsResult {
-  public readonly operations: Operation[]
-  public readonly reason?: LiveValidationError
-  public constructor(operations: Operation[], reason: undefined | LiveValidationError) {
+  public constructor(
+    public readonly operations: Operation[],
+    public readonly resourceProvider: string,
+    public readonly apiVersion: string,
+    public readonly reason?: undefined | LiveValidationError
+  ) {
     this.operations = operations || []
     if (reason) {
       this.reason = reason